requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
Jinja2==3.1.3
boto3>=1.34.0
